requirements.txt
streamlit
tensorflow>=2.10
numpy
pillow
opencv-python-headless
matplotlib
seaborn
scikit-learn
plotly
